Failed to get feature working, need to refactor...
Can't figure out how the async stuff is working between the
readline module and the API call to Google Calendar, so I'm going
to refactor the project to simply display my Calendar or Email
from a passed in command-line option. This will remove the
dependency on readline and should solve the async issues.